+++ agents/logsuploader/samples/java.txt
@@ -1,0 +1,11 @@
+{"time":"2020-07-24T00:22:30.259Z","level":"INFO","logger":"applicationinsights.extension.diagnostics","message":"Application Insights Codeless Agent Attach Successful","properties":{"operation":"Startup","siteName":"unknown","ikey":"0ea304c8-8ce6-4626-8515-5407d2c7bdd1","extensionVersion":"unknown","sdkVersion":"3.0.0-PREVIEW.5","
+subscriptionId":"unknown"}}
+sdfsdfsd
+fdsfsd
+fsdfsdfsd
+vcvc
+dasdasda
+sgdgdfg
+gdfgdgd
+aaa
+bbbb
