+++ helm/values.yaml
@@ -8,10 +8,8 @@ app:
   name: "app-monitoring-webhook"
   environment: "test"
   owner : "Microsoft"
-  image: "mcr.microsoft.com/applicationinsights/codeless-attach/mutating-webhook:test7_20_1"
-  #agents: "gearamaaks.azurecr.io/public/applicationinsights/codeless-attach/mutating-webhook-agents:v31"
-  #agents: "mcr.microsoft.com/applicationinsights/codeless-attach/mutating-webhook-agents:preview2833"
-  agents: "gearamaaks.azurecr.io/public/applicationinsights/codeless-attach/ai_agents:test7_20_3"
+  image: "mcr.microsoft.com/applicationinsights/codeless-attach/mutating-webhook:test7_23_7"
+  agents: "gearamaaks.azurecr.io/public/applicationinsights/codeless-attach/ai_agents:test7_23_7"
   caBundle: "baseBundle"
   iKey: "target ikey2"
   telemetryIkey: "320dcf98-173f-429b-ab39-df8b4951fb94"
